+++ lib/src/search_app_bar_page/controller/utils/filters/functions_filters.dart
@@ -2,6 +2,7 @@ import 'package:diacritic/diacritic.dart';
 import 'package:flutter/material.dart';
 
 typedef Filter<T> = bool Function(T test, String? query);
+typedef SortList<T> = int Function(T a, T b);
 
 //typedef StringFilter<T> = String Function(T test);